# api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AddManual" : {
        "description" : "Mark one transaction as a recurring payment (or recurring income)",
        "properties" : {
          "cadence" : {
            "enum" : [ "DAILY", "WEEKLY", "BIWEEKLY", "MONTHLY", "QUARTERLY", "YEARLY" ],
            "type" : "string"
          },
          "name" : {
            "description" : "Defaults to the merchant's name",
            "type" : "string"
          },
          "transactionId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "cadence", "transactionId" ],
        "type" : "object"
      },
...
      "Category" : {
        "properties" : {
          "code" : {
            "example" : "GROCERIES",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "description" : "Only SPEND counts towards spent",
            "enum" : [ "SPEND", "INCOME", "TRANSFER" ],
            "type" : "string"
          },
          "name" : {
            "example" : "Groceries",
            "type" : "string"
          },
          "parentId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "code", "id", "kind", "name" ],
        "type" : "object"
      },
...
      "CategoryDetail" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "averageMinor" : {
            "description" : "Average over the trend months that have any data; absent when none has",
            "format" : "int64",
            "type" : "integer"
          },
          "category" : {
            "$ref" : "#/components/schemas/Category"
          },
          "currency" : {
            "example" : "RON",
            "type" : "string"
          },
          "merchants" : {
            "description" : "This month's variable spending by merchant (recurring charges excluded), largest first",
            "items" : {
              "$ref" : "#/components/schemas/MerchantAmount"
            },
            "type" : "array"
          },
          "month" : {
            "example" : "2026-03",
            "type" : "string"
          },
          "recurring" : {
            "description" : "This month's recurring payments, largest first",
            "items" : {
              "$ref" : "#/components/schemas/RecurringAmount"
            },
            "type" : "array"
          },
          "recurringMinor" : {
            "description" : "This month's recurring charges, total",
            "format" : "int64",
            "type" : "integer"
          },
          "trend" : {
            "description" : "The 12 months ending with `month`, oldest first",
            "items" : {
              "$ref" : "#/components/schemas/MonthAmount"
            },
            "type" : "array"
          }
        },
        "required" : [ "amountMinor", "category", "currency", "merchants", "month", "recurring", "recurringMinor", "trend" ],
        "type" : "object"
      },
...
      "Confirm" : {
        "description" : "Corrections made while confirming; omitted fields keep the detected value",
        "properties" : {
          "cadence" : {
            "enum" : [ "DAILY", "WEEKLY", "BIWEEKLY", "MONTHLY", "QUARTERLY", "YEARLY" ],
            "type" : "string"
          },
          "expectedAmountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Coverage" : {
        "properties" : {
          "accountId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "accountName" : {
            "type" : "string"
          },
          "from" : {
            "format" : "date",
            "type" : "string"
          },
          "monthly" : {
            "description" : "3+ months: monthly payments detectable",
            "type" : "boolean"
          },
          "months" : {
            "description" : "Calendar months from first to last date",
            "format" : "int32",
            "type" : "integer"
          },
          "to" : {
            "format" : "date",
            "type" : "string"
          },
          "yearly" : {
            "description" : "13+ months: yearly payments detectable",
            "type" : "boolean"
          }
        },
        "required" : [ "accountId", "accountName", "monthly", "months", "yearly" ],
        "type" : "object"
      },
...
      "Group" : {
        "properties" : {
          "items" : {
            "description" : "Highest monthly equivalent first",
            "items" : {
              "$ref" : "#/components/schemas/Item"
            },
            "type" : "array"
          },
          "kind" : {
            "enum" : [ "SUBSCRIPTIONS", "BILLS", "INCOME" ],
            "type" : "string"
          },
          "monthlyMinor" : {
            "description" : "Counted items only",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "items", "kind", "monthlyMinor" ],
        "type" : "object"
      },
...
      "MerchantAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "merchantId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "transactionCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "amountMinor", "merchantId", "name", "transactionCount" ],
        "type" : "object"
      },
...
      "Merge" : {
        "properties" : {
          "into" : {
            "description" : "The recurring payment it is the same as",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "into" ],
        "type" : "object"
      },
      "MonthAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "hasData" : {
            "description" : "Whether any transaction exists in this month",
            "type" : "boolean"
          },
          "month" : {
            "example" : "2026-03",
            "type" : "string"
          }
        },
        "required" : [ "amountMinor", "hasData", "month" ],
        "type" : "object"
      },
...
      "RecurringAmount" : {
        "properties" : {
          "amountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "cadence" : {
            "enum" : [ "DAILY", "WEEKLY", "BIWEEKLY", "MONTHLY", "QUARTERLY", "YEARLY" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "state" : {
            "description" : "CONFIRMED or ENDED",
            "type" : "string"
          },
          "subscriptionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "transactionCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "amountMinor", "cadence", "name", "state", "subscriptionId", "transactionCount" ],
        "type" : "object"
      },
      "RecurringOverview" : {
        "properties" : {
          "countedCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "coverage" : {
            "description" : "Per account: how much history detection has",
            "items" : {
              "$ref" : "#/components/schemas/Coverage"
            },
            "type" : "array"
          },
          "currency" : {
            "type" : "string"
          },
          "groups" : {
            "description" : "Non-empty groups, subscriptions first",
            "items" : {
              "$ref" : "#/components/schemas/Group"
            },
            "type" : "array"
          },
          "incomeMonthlyMinor" : {
            "description" : "Recurring income per month: the INCOME group's counted items (a salary in two parts is two)",
            "format" : "int64",
            "type" : "integer"
          },
          "month" : {
            "description" : "The month viewed (YYYY-MM); absent = as of today",
            "type" : "string"
          },
          "monthlyMinor" : {
            "description" : "Committed per month: counted items only",
            "format" : "int64",
            "type" : "integer"
          },
          "standingMonthlyMinor" : {
            "description" : "Standing transfers per month: active ones only",
            "format" : "int64",
            "type" : "integer"
          },
          "standingTransfers" : {
            "description" : "Recurring transfers to own accounts or savings: shown, never part of the totals",
            "items" : {
              "$ref" : "#/components/schemas/StandingTransfer"
            },
            "type" : "array"
          },
          "suggestionCount" : {
            "description" : "Suggestions waiting in the review inbox",
            "format" : "int32",
            "type" : "integer"
          },
          "yearlyMinor" : {
            "description" : "Committed per year: counted items only",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "countedCount", "coverage", "currency", "groups", "incomeMonthlyMinor", "monthlyMinor", "standingMonthlyMinor", "standingTransfers", "suggestionCount", "yearlyMinor" ],
        "type" : "object"
      },
      "Reminder" : {
        "description" : "Days before the next charge, 1–14; absent or null turns the reminder off",
        "properties" : {
          "daysBefore" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Rename" : {
        "properties" : {
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
...
      "Split" : {
        "properties" : {
          "atMinor" : {
            "description" : "Charges below this amount (positive minor units) are one plan, the rest the other",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "atMinor" ],
        "type" : "object"
      },
...
      "Subscription" : {
        "properties" : {
          "accountId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "amountKind" : {
            "enum" : [ "FIXED", "VARIABLE" ],
            "type" : "string"
          },
          "anchorDay" : {
            "description" : "Monthly, quarterly, yearly: day of month the charge is due (clamped in shorter months); weekly, bi-weekly: ISO weekday (1 = Monday)",
            "format" : "int32",
            "type" : "integer"
          },
          "anchorMonth" : {
            "description" : "Yearly: month of year. Quarterly: the cycle's first month, 1–3 (1 = Jan/Apr/Jul/Oct). Bi-weekly: which of the alternating weeks, 1 or 2",
            "format" : "int32",
            "type" : "integer"
          },
          "bandMaxMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "bandMinMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "cadence" : {
            "enum" : [ "DAILY", "WEEKLY", "BIWEEKLY", "MONTHLY", "QUARTERLY", "YEARLY" ],
            "type" : "string"
          },
          "confidence" : {
            "description" : "Detector confidence, 0–1",
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "direction" : {
            "description" : "OUT: a recurring payment; IN: recurring income",
            "enum" : [ "IN", "OUT" ],
            "type" : "string"
          },
          "endedBy" : {
            "enum" : [ "USER", "SYSTEM" ],
            "type" : "string"
          },
          "expectedAmountMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "firstSeen" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lastSeen" : {
            "format" : "date",
            "type" : "string"
          },
          "merchantId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "nextExpectedDate" : {
            "format" : "date",
            "type" : "string"
          },
          "remindDaysBefore" : {
            "description" : "Remind this many days before the next charge; absent = no reminder",
            "format" : "int32",
            "type" : "integer"
          },
          "state" : {
            "enum" : [ "PROPOSED", "CONFIRMED", "REJECTED", "ENDED" ],
            "type" : "string"
          },
          "toleranceMinor" : {
            "description" : "A charge within expected ± tolerance matches",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "accountId", "amountKind", "bandMaxMinor", "bandMinMinor", "cadence", "confidence", "currency", "direction", "expectedAmountMinor", "firstSeen", "id", "lastSeen", "merchantId", "name", "state", "toleranceMinor" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "Outflow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/insights/categories/{id}" : {
      "get" : {
        "operationId" : "category",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "YYYY-MM",
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "ISO currency; absent = the main currency of the accounts (most transactions)",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Account ids to include (accounts filter); absent = all accounts",
          "in" : "query",
          "name" : "accounts",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "insights" ]
      }
    },
...
    "/api/subscriptions" : {
      "get" : {
        "operationId" : "overview",
        "parameters" : [ {
          "description" : "YYYY-MM; absent = as of today",
          "in" : "query",
          "name" : "month",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "ISO currency; absent = the main currency of the accounts (most transactions)",
          "in" : "query",
          "name" : "currency",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Account ids to include (accounts filter); absent = all accounts",
          "in" : "query",
          "name" : "accounts",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecurringOverview"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      },
      "post" : {
        "operationId" : "addManual",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddManual"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subscription"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
...
    "/api/subscriptions/{id}" : {
      "patch" : {
        "operationId" : "rename",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Rename"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subscription"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
    "/api/subscriptions/{id}/confirm" : {
      "post" : {
        "operationId" : "confirm",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Confirm"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subscription"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
    "/api/subscriptions/{id}/end" : {
      "post" : {
        "operationId" : "end",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subscription"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
    "/api/subscriptions/{id}/merge" : {
      "post" : {
        "operationId" : "merge",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Merge"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subscription"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
    "/api/subscriptions/{id}/reject" : {
      "post" : {
        "operationId" : "reject",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subscription"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
    "/api/subscriptions/{id}/reminder" : {
      "put" : {
        "operationId" : "reminder",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Reminder"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subscription"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
    "/api/subscriptions/{id}/split" : {
      "post" : {
        "operationId" : "split",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Split"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Subscription"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "subscriptions" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}